+++ src-tauri/src/kernel/model/artist.rs
@@ -18,8 +18,8 @@ pub struct NewArtist {
 
 #[derive(new, Debug)]
 pub struct SearchAlsoUsingWorkArtist {
-    pub limit: u8,
-    pub offset: u8,
+    pub limit: u16,
+    pub offset: u16,
     pub sort_col: String,
     pub sort_desc: bool,
     pub text: String,
