+++ src-tauri/src/kernel/repository/tag.rs
@@ -10,6 +10,6 @@ pub trait TagRepository {
     async fn find_by_ids(&self, ids: &Vec<Id<Tag>>) -> anyhow::Result<Vec<Tag>>;
     async fn insert(&self, source: NewTag) -> anyhow::Result<()>;
     async fn find_by_name(&self, name: String) -> anyhow::Result<Option<Tag>>;
-    async fn select(&self, limit: u8) -> anyhow::Result<Vec<Tag>>;
+    async fn select(&self, limit: u16) -> anyhow::Result<Vec<Tag>>;
     async fn search_by_name(&self, name: &str) -> anyhow::Result<Vec<Tag>>;
 }
